requirements.txt
Pillow>9,<10
numpy>1.23,<2
rawpy>0.17,<0.18
